# Job details rendering performance optimization