--- a/templates/webapi/test/details.html.ep
+++ b/templates/webapi/test/details.html.ep
@@ -13,4 +13,3 @@
         <label for="details-only-failed-filter">Show only failures</label>
     </div>
 </form>
-%= include 'test/module_table'