Server-side template rendering is extremely slow with large number of test modules (100+) in a job. Send module results to client as JSON and render HTML on client side using JavaScript. This makes rendering almost 30 times faster.

Currently, rendering of LTP syscalls job results (1112 modules) into HTML in OpenQA takes over 16 seconds. 3 seconds are spent reading and parsing data from storage, 13 seconds are spent rendering the HTML template. With this PR, HTML rendering takes only 700ms on my machine.

The JavaScript generates HTML functionally equivalent to the `test/module_table` template. The only significant change I did was replacing the `<img>` elements for terminal and audio boxes with CSS-defined equivalent `<span>` elements. This little change sped up rendering of syscalls jobs from 2.5 seconds to 700 milliseconds.

There will be a second part of this PR which will optimize `read_test_modules()` and possibly speed up JSON generation by another 2 seconds. But I have another task to take care of first and I don't mind opening a separate PR for part 2.