@pevik 2020-05-12
Thanks for a speedup. BTW which browser did you test? Could you post the result (verification)?