@pevik 2020-05-12
BTW openQA use jQuery (v3.3.1 https://openqa.suse.de/asset/8b091c3fbc/bootstrap.js), code might be smaller, if used, particularly for DOM manipulation.