@mdoucha 2020-05-12
> Thanks for a speedup. BTW which browser did you test? Could you post the result (verification)?

I've tested it in Firefox and Chromium. Unfortunately, I don't have a publicly accessible OpenQA instance where I could show the results.