@mdoucha 2020-05-12
> I _like_ that you don't use jQuery. The library is old and projects like Bootstrap are moving away from it. Therefore it is a good choice to avoid using it in new code - especially for the sake of saving a few lines of code. So please do _not_ follow @pevik suggestion of using it more.

Don't worry, I have no such plans. jQuery doesn't really offer any decent tools for rendering a full element tree. But it's nice when you need to manipulate an existing one, which is not the case here.

> Speaking of tests: I'd also like to test this locally. Can you point me to a specific job I can clone where the performance difference is especially apparent?

[ltp_openposix](https://openqa.suse.de/tests/4196768): 1606 modules, 88 seconds rendering time right now on OSD (should be 25 seconds on my local instance before patch).
[ltp_syscalls](https://openqa.suse.de/tests/4196749): 1112 modules, 83 seconds rendering time (16 seconds on my local instance before patch).

> One issue which comes to my mind is the "External results" table. It is so far only rendered via a template and it doesn't seem you've adjusted that. It makes most sense to convert this table to JSON as well (at some point) and so it would be good to keep that already in mind. Both views could possible shared code.

The template for External Results tab is a lot different so I'd have to write a new `renderTable()` and `renderRow()` for it either way. 90% of the JSON data needed for Details tab is also irrelevant for External Results so it's better to handle these separately.