@mdoucha 2020-05-19
New version. Changelog:
- code style fixes
- job_details.js renamed to render.js
- use `include_branding()` in JSON generation
- unit test fixes
- added ID to each module row for later use