@mdoucha 2020-05-19
> I've tested it locally. While the test is running everything is fine and modules are added as the test progresses. However, when the test status changes to `done` almost all test modules are removed again from the page and the preview container is missing as well:

I've tried watching live updates of the module list in Firefox and Chromium and I can't reproduce the issue on my instance. I suspect that on your instance, it was caused by browser cache. Live updates still use HTML rendered on the server side and only the final update would switch to client-side rendered JSON. Can you restart the job and see if it happens again? Any error messages from web console would also be nice.