@mdoucha 2020-05-19
Small correction, running jobs don't use my code at all. Live updates all the way to the end use server-side rendering through `module_components_ajax` (`Controller::Test::module_components()`) and the client DOM is updated by `updateTestStatus()` in `running.js`.