@mdoucha 2020-05-19
One last fix for a small bug in generating links for duplicate modules (their name includes `#`):

    diff --git a/assets/javascripts/render.js b/assets/javascripts/render.js
    index de27bbdbf..1680e0c36 100644
    --- a/assets/javascripts/render.js
    +++ b/assets/javascripts/render.js
    @@ -82,7 +82,8 @@ function renderModuleRow(module, snippets) {
             }));
         }
     
    -    let src_url = renderTemplate(snippets.src_url, {MODULE: module.name});
    +    let src_url = renderTemplate(snippets.src_url,
    +        {MODULE: encodeURIComponent(module.name)});
         let component = E('td', [
             E('div', [E('a', [module.name], {href: src_url})]),
             E('div', flags, {'class': 'flags'})
    @@ -95,7 +96,7 @@ function renderModuleRow(module, snippets) {
             let step = module.details[idx];
             let title = step.display_title;
             let href = '#step/' + module.name + '/' + step.num;
    -        let tplargs = {MODULE: module.name, STEP: step.num};
    +        let tplargs = {MODULE: encodeURIComponent(module.name), STEP: step.num};
             let alt = '';
     
             if (step.name) {
    @@ -104,7 +105,8 @@ function renderModuleRow(module, snippets) {
     
             if (step.is_parser_text_result) {
                 let elem = E('span', [], {'class': 'step_actions'});
    -            elem.innerHTML = renderTemplate(snippets.bug_actions, tplargs);
    +            elem.innerHTML = renderTemplate(snippets.bug_actions,
    +                {MODULE: module.name, STEP: step.num});
                 elem = E('span', [elem, step.text_data],
                     {'class': 'resborder ' + step.resborder});
                 elem = E('span', [elem], {title: title, 'data-href': href,