@Martchus 2020-05-20
I've just checked the added/changed JS files with `jshint` and it didn't complain about something.

@okurz OSD has been deployed and the points mentioned in your previous review have been addressed. So can we merge this PR now?